feat: granularize /unreserve
